requirements.txt
Flask==2.2.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
Jinja2==3.1.2
pyflakes==3.0.1
python-dotenv==1.0.0
SQLAlchemy==2.0.6
Werkzeug==2.2.3
WTForms==3.0.1
